refactor: code cleanup activity node creation logic
